Python-Code/requirements.txt
plotly==4.14.3
dash==1.19.0
dash_html_components==1.1.2
dash_core_components==1.15.0
paho_mqtt==1.5.1
